requirements.txt
flask==2.2.2
kafka-python==2.0.2
flask-kafka==0.0.3
elasticsearch==8.3.3